# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  subcategory: string;
  /** @nullable */
  description?: string | null;
  basePrice: number;
  basePricePkr: number;
  availableColors: string[];
  availableFabrics: string[];
  availableGsm: number[];
  minOrderQty: number;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  createdAt: string;
}

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  startingPricePkr: number;
  countriesServed: number;
}

export interface CategoryCount {
  category: string;
  count: number;
  subcategories: string[];
}

export interface Design {
  id: number;
  name: string;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  designData: string;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  prompt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDesignBody {
  name: string;
  /** @nullable */
  productId?: number | null;
  designData: string;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  prompt?: string | null;
}

export interface QuoteRequest {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  items: string;
  totalUnits: number;
  estimatedTotal: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateQuoteRequestBody {
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  items: string;
  totalUnits: number;
  estimatedTotal: number;
  /** @nullable */
  notes?: string | null;
}

export interface QuoteLineItem {
  productId: number;
  quantity: number;
  basePricePkr: number;
}

export interface CalculateQuoteBody {
  items: QuoteLineItem[];
}

export interface CalculatedLineItem {
  productId: number;
  quantity: number;
  unitPricePkr: number;
  subtotalPkr: number;
}

export interface QuoteCalculation {
  items: CalculatedLineItem[];
  totalUnits: number;
  subtotalPkr: number;
  discountPercent: number;
  discountAmountPkr: number;
  totalPkr: number;
  discountTier: string;
  /** @nullable */
  nextTierUnits?: number | null;
  /** @nullable */
  nextTierDiscount?: number | null;
  productionDays: number;
  requiresCustomPricing: boolean;
}

export type ListProductsParams = {
  category?: string;
  subcategory?: string;
  fabric?: string;
  limit?: number;
  offset?: number;
};