# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const listProductsQueryLimitDefault = 50;
export const listProductsQueryOffsetDefault = 0;

export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  subcategory: zod.coerce.string().optional(),
  fabric: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listProductsQueryLimitDefault),
  offset: zod.coerce.number().default(listProductsQueryOffsetDefault),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  subcategory: zod.string(),
  description: zod.string().nullish(),
  basePrice: zod.number(),
  basePricePkr: zod.number(),
  availableColors: zod.array(zod.string()),
  availableFabrics: zod.array(zod.string()),
  availableGsm: zod.array(zod.number()),
  minOrderQty: zod.number(),
  imageUrl: zod.string().nullish(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Get a product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  subcategory: zod.string(),
  description: zod.string().nullish(),
  basePrice: zod.number(),
  basePricePkr: zod.number(),
  availableColors: zod.array(zod.string()),
  availableFabrics: zod.array(zod.string()),
  availableGsm: zod.array(zod.number()),
  minOrderQty: zod.number(),
  imageUrl: zod.string().nullish(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get catalog summary stats
 */
export const GetCatalogSummaryResponse = zod.object({
  totalProducts: zod.number(),
  totalCategories: zod.number(),
  startingPricePkr: zod.number(),
  countriesServed: zod.number(),
});

/**
 * @summary List all product categories with counts
 */
export const ListCategoriesResponseItem = zod.object({
  category: zod.string(),
  count: zod.number(),
  subcategories: zod.array(zod.string()),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get featured/popular products for homepage
 */
export const GetFeaturedProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  subcategory: zod.string(),
  description: zod.string().nullish(),
  basePrice: zod.number(),
  basePricePkr: zod.number(),
  availableColors: zod.array(zod.string()),
  availableFabrics: zod.array(zod.string()),
  availableGsm: zod.array(zod.number()),
  minOrderQty: zod.number(),
  imageUrl: zod.string().nullish(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});
export const GetFeaturedProductsResponse = zod.array(
  GetFeaturedProductsResponseItem,
);

/**
 * @summary List saved designs
 */
export const ListDesignsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  productId: zod.number().nullish(),
  productName: zod.string().nullish(),
  designData: zod.string(),
  previewUrl: zod.string().nullish(),
  prompt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListDesignsResponse = zod.array(ListDesignsResponseItem);

/**
 * @summary Save a new design
 */
export const CreateDesignBody = zod.object({
  name: zod.string(),
  productId: zod.number().nullish(),
  designData: zod.string(),
  previewUrl: zod.string().nullish(),
  prompt: zod.string().nullish(),
});

/**
 * @summary Get a saved design
 */
export const GetDesignParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDesignResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  productId: zod.number().nullish(),
  productName: zod.string().nullish(),
  designData: zod.string(),
  previewUrl: zod.string().nullish(),
  prompt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a saved design
 */
export const DeleteDesignParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Submit a bulk order quote request
 */
export const CreateQuoteRequestBody = zod.object({
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.string(),
  totalUnits: zod.number(),
  estimatedTotal: zod.number(),
  notes: zod.string().nullish(),
});

/**
 * @summary Calculate pricing for a quote (no DB write)
 */
export const CalculateQuoteBody = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.number(),
      quantity: zod.number(),
      basePricePkr: zod.number(),
    }),
  ),
});

export const CalculateQuoteResponse = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.number(),
      quantity: zod.number(),
      unitPricePkr: zod.number(),
      subtotalPkr: zod.number(),
    }),
  ),
  totalUnits: zod.number(),
  subtotalPkr: zod.number(),
  discountPercent: zod.number(),
  discountAmountPkr: zod.number(),
  totalPkr: zod.number(),
  discountTier: zod.string(),
  nextTierUnits: zod.number().nullish(),
  nextTierDiscount: zod.number().nullish(),
  productionDays: zod.number(),
  requiresCustomPricing: zod.boolean(),
});