# lib/api-zod/src/generated/types/calculateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteLineItem } from "./quoteLineItem";

export interface CalculateQuoteBody {
  items: QuoteLineItem[];
}