# lib/api-zod/src/generated/types/calculatedLineItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CalculatedLineItem {
  productId: number;
  quantity: number;
  unitPricePkr: number;
  subtotalPkr: number;
}