# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryCount {
  category: string;
  count: number;
  subcategories: string[];
}