# lib/api-zod/src/generated/types/createDesignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDesignBody {
  name: string;
  /** @nullable */
  productId?: number | null;
  designData: string;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  prompt?: string | null;
}