# lib/api-zod/src/generated/types/design.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Design {
  id: number;
  name: string;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  designData: string;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  prompt?: string | null;
  createdAt: string;
  updatedAt: string;
}