# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  subcategory: string;
  /** @nullable */
  description?: string | null;
  basePrice: number;
  basePricePkr: number;
  availableColors: string[];
  availableFabrics: string[];
  availableGsm: number[];
  minOrderQty: number;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  createdAt: string;
}