# lib/api-zod/src/generated/types/quoteCalculation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalculatedLineItem } from "./calculatedLineItem";

export interface QuoteCalculation {
  items: CalculatedLineItem[];
  totalUnits: number;
  subtotalPkr: number;
  discountPercent: number;
  discountAmountPkr: number;
  totalPkr: number;
  discountTier: string;
  /** @nullable */
  nextTierUnits?: number | null;
  /** @nullable */
  nextTierDiscount?: number | null;
  productionDays: number;
  requiresCustomPricing: boolean;
}