# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signitive Enterprises API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteRequest {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  items: string;
  totalUnits: number;
  estimatedTotal: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}