# lib/db/src/schema/designs.ts
import { pgTable, text, serial, timestamp, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const designsTable = pgTable("designs", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  productId: integer("product_id"),
  productName: text("product_name"),
  designData: text("design_data").notNull(),
  previewUrl: text("preview_url"),
  prompt: text("prompt"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertDesignSchema = createInsertSchema(designsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertDesign = z.infer<typeof insertDesignSchema>;
export type Design = typeof designsTable.$inferSelect;